requirements.txt
django==1.5.8
pillow
mysql-python
django-bootstrap-form
django-tastypie